WakeLock, Vibrations, Webmanifest + styles
- screen lock on supported devices with WakeLock API, to prevent the Device
from going to sleep and turn down the display.

- Long and short 'Track gate style' count down signals with vibrations
Activated when next interval is about to start. For mobile devices.

- pwa web manifest with basic icons

- chrome background color added to the quick connection screen for dark theme
to bring some proper app feel. Background is configured to the same color
for all devices/browser that support that kind of thing through webmanifest.

- fixed stop watch under counting the last interval index

- stop watch gets into normal count UP mode for the current interval when
the workout is over, so if the workout was too short the User can do more
Intervals with manual mode.

- screens are grouped into blocks of pages