moving to 7 zone power system
Sweet Spot is the single most important training zone and in the Zwift
power zone system it does not have an individual spot and is split
between green and mostly yellow zone. The colors, functions and html
are reworked to the new system where Sweet Spot will be green and Threshold
will be Yellow. Tempo zone being the junk mile zone is assaigned a gray
color.