xf improvements
- xf is class containing a tiny reactive library:
  - a Proxy named db is used for general app state
  - store -> creates a Proxy with a set trap
  - reg -> subscribes to events and gets db as second param
    it can write to db state
  - sub -> subscribes to state changes, ui or other custom app events
  - dispatch -> emits a custom event
  - initial app state is described in an Object and passed to initDB.

- started to remove local state in favour of global.

Note: I wanted this to be a frameworkless, dependancy free project,
but I am slowly converging to the familiar reactive framework design.
I'll at least try to keep it simple and obvious without any pointless
complexity.
Having said that, file structure and naming is a mess, I'll fix that
as soon as I converge on a more solid design.