switching data sources between Controllable and Power Meter
Use case:
The User wants to have a Trainer connected for resistance/simulation control,
but read power data from a specialized power meter.

User can connect a Controllable (Trainer) and Power Meter, and
pick the source for power, cadence, and speed data between them,
by clicking on their corresponding values in Settings.
When a source is active it has white text color,
when inactive it is gray.

power, cadence and speed default to Controllable, until a power meter
is connected, then it becomes the selected source for power, but not
cadence and speed.