fix MessageTransformer and ANT+ channel
- Message Transformer goes recursively on the current chunk,
when sync byte is encountered it reads length of the message (next byte)
and copies the message into an accumulator.
If the last message is not full length it stores it into a leftover
buffer for the next chunk.
If first byte is not sync and there is leftover it concats them.

- fixing ANT channel
  - writeWithResponse(), where the resolve reject functions were switched,
  and it always rejected the received response, thus triggering the timeout.
  - onData(), where an open channel check prevented response messages from
  being read when configuring a channel that is still not open.

- fixing ANT+ request popup found devices list not being cleared
after cancel or pair.